Update remaining time once a second + Bug fixed on huge files transfer
+ Now the remaining time is updated only once a secondo, avoiding useless updates and making the whole thing more stable.
+ There was an overflow bug on big files transfer: the offset in the FileIterator was an int and so was the SentByte in the Job. Making both long the overflow doesn't appear anymore. A similar correction was made on 'sentFromStartTime' in ListedJob.

+++ FileShare/FileShare/Job.cs
@@ -44,11 +44,11 @@ namespace FileTransfer
             }
         }
 
-        private int sentByte;
+        private long sentByte;
 
         public event PropertyChangedEventHandler PropertyChanged;
 
-        public int SentByte {
+        public long SentByte {
             get {
                 return sentByte;
             }
